pyproject.toml
[project]
name = "python-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.28.0",
    "fastapi>=0.100.0",
    "git-filter-repo>=2.47.0",
    "instructor>=0.4.0",
    "litellm>=1.10.0",
    "loguru>=0.7.3",
    "mangum>=0.17.0",
    "pillow>=10.0.0",
    "pydantic>=2.4.0",
    "pypdf>=3.15.0",
    "pytest>=7.3.1",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
]
